Discard SWBParser. Not needed.
