+++ js/SWBDatasource.js
@@ -1,4 +1,7 @@
 import * as THREE from 'three'
+import {openH5File} from 'hdf5-indexed-reader'
+import {FileUtils} from 'igv-utils'
+import {SpacewalkGlobals} from './app.js'
 import DataSourceBase from './dataSourceBase.js'
 import {hideGlobalSpinner, showGlobalSpinner} from "./utils";
 import {createBoundingBoxWithFlatXYZList} from "./math.js"
@@ -8,6 +11,18 @@ class SWBDatasource extends DataSourceBase {
     constructor() {
         super()
         this.currentXYZList = undefined
+    }
+
+    async parse(path, datasource) {
+
+        SpacewalkGlobals.url = false === FileUtils.isFilePath(path) ? path : undefined
+
+        const hdf5 = await openH5File(FileUtils.isFilePath(path) ? { file:path } : { url: path })
+
+        const { sample, genomeAssembly } = await datasource.initialize(hdf5)
+
+        return { sample, genomeAssembly }
+
     }
 
     async initialize(hdf5) {
